+++ Cargo.toml
@@ -9,10 +9,10 @@ edition = "2018"
 [dependencies]
 ark-ff = { version = "^0.3.0", default-features = false }
 ark-std = { version = "^0.3.0", default-features = false }
-ark-relations = { version = "^0.3.0", default-features = false }
 ark-r1cs-std = { version = "^0.3.0", default-features = false}
 ark-sponge = { version = "^0.3.0", default-features = false }
 ark-poly = { version = "0.3.0", default-features = false }
+ark-relations = { version = "^0.3.0", default-features = false, optional = true}
 tracing = { version = "0.1", default-features = false, features = [ "attributes" ], optional = true}
 
 
@@ -22,4 +22,4 @@ ark-test-curves = { version = "^0.3.0", default-features = false, features = ["b
 [features]
 default = ["std"]
 std = ["ark-ff/std", "ark-std/std", "ark-relations/std", "ark-r1cs-std/std", "ark-sponge/std", "ark-poly/std"]
-r1cs = ["ark-sponge/r1cs", "tracing"]
+r1cs = ["ark-sponge/r1cs", "tracing", "ark-relations"]
